move `Inet` object out of ent schema package
